+++ server/component/userUpdate/UserUpdateView.php
@@ -105,6 +105,46 @@ class UserUpdateView extends BaseView
     }
 
     /**
+     * Render the form to send activation email to a user.
+     */
+    private function output_form_activation_email()
+    {
+        $form = new BaseStyleComponent("card", array(
+            "css" => "mb-3",
+            "is_expanded" => true,
+            "is_collapsible" => false,
+            "title" => "Send Activation Email",
+            "type" => "info",
+            "children" => array(
+                new BaseStyleComponent("plaintext", array(
+                    "text" => "The user will recive an activation email",
+                    "is_paragraph" => true,
+                )),
+                new BaseStyleComponent("form", array(
+                    "label" => "Send Email",
+                    "url" => $this->model->get_link_url("userUpdate",
+                        array(
+                            "uid" => $this->selected_user['id'],
+                            "mode" => "activation_email",
+                        )
+                    ),
+                    "type" => "info",
+                    "url_cancel" => $this->model->get_link_url("userSelect",
+                        array("uid" => $this->selected_user['id'])),
+                    "children" => array(
+                        new BaseStyleComponent("input", array(
+                            "type_input" => "hidden",
+                            "name" => "activation_email",
+                            "value" => $this->selected_user['email']
+                        )),
+                    )
+                )),
+            )
+        ));
+        $form->output_content();
+    }
+
+    /**
      * Render the form to block a user.
      */
     private function output_form_block()
@@ -269,6 +309,8 @@ class UserUpdateView extends BaseView
             require __DIR__ . "/tpl_success_block.php";
         else if($type === "unblock")
             require __DIR__ . "/tpl_success_unblock.php";
+         else if($type === "activation_email")
+            require __DIR__ . "/tpl_success_activation_email.php";
         else if($type === "clean")
             require __DIR__ . "/tpl_success_clean.php";
         else if($type === "add_group")
@@ -302,6 +344,13 @@ class UserUpdateView extends BaseView
         $url_user = $this->model->get_link_url("userSelect",
             array("uid" => $this->selected_user['id']));
         $url_users = $this->model->get_link_url("userSelect");
+        if($this->mode == "activation_email")
+        {
+            if($this->controller->has_succeeded())
+                require __DIR__ . "/tpl_success.php";
+            else
+                require __DIR__ . "/tpl_activation_email.php";
+        }
         if($this->mode == "block")
         {
             if($this->controller->has_succeeded())