+++ server/component/userUpdate/tpl_success_activation_email.php
@@ -1,0 +1,6 @@
+<?php
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
+?>
+<p>Activation email was successfully sent to <code><?php echo $this->selected_user['email']?></code>.</p>